www/faq.md: Provide an example for padding ROMs
Plenty of ROMs provided in the 20160907 release do not have
configurations for each possible flash chip size (that the board
supports). As a result, padding will be necessary in those cases
where there is a difference in size between the ROM and flash chip.

Providing an example of how to do this in the FAQ makes it easy to
link to as necessary.